chore: Adiciona logs globais e define retornos HTTP
