+++ src/arquivos/controllers/arquivos.controller.ts
@@ -16,10 +16,10 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import { ApiConsumes, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Permissions } from 'src/authentication/decorators';
 import { arquivos } from 'src/common/resources';
-import { permissoesArquivos } from 'src/common/resources/permissoes.resources';
+import { permissoesArquivos } from 'src/common/resources';
 import { arquivoPdfValidator } from 'src/common/validators';
 import { ArquivoEntity } from '../entities';
-import { ArquivoInclusaoModel } from '../models';
+import { ArquivoModel } from '../models';
 import { ArquivoConteudoSerializer, ArquivoSerializer } from '../serializers';
 import { ArquivosService } from '../services';
 
@@ -33,6 +33,7 @@ export class ArquivosController {
   ) {}
 
   @Get(':id')
+  @HttpCode(HttpStatus.OK)
   @Permissions(permissoesArquivos.consulta)
   @ApiOperation({ summary: arquivos.consulta })
   consulta(@Param('id') id: string): Promise<ArquivoEntity> {
@@ -40,6 +41,7 @@ export class ArquivosController {
   }
 
   @Get(':id/download')
+  @HttpCode(HttpStatus.OK)
   @Header('Content-Disposition', 'attachment; filename=download.pdf')
   @Permissions(permissoesArquivos.consulta)
   @ApiOperation({ summary: arquivos.download })
@@ -54,7 +56,7 @@ export class ArquivosController {
   @ApiOperation({ summary: arquivos.inclui })
   @UseInterceptors(FileInterceptor('conteudo'))
   async inclui(
-    @Body() model: ArquivoInclusaoModel,
+    @Body() model: ArquivoModel,
     @UploadedFile(arquivoPdfValidator)
     conteudo: Express.Multer.File,
   ): Promise<Pick<ArquivoEntity, 'id'>> {
