+++ src/arquivos/serializers/arquivo.serializer.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { BaseSerializer } from 'src/common/serializers/base.serializer';
-import { ArquivoInclusaoModel } from '../models';
+import { ArquivoModel } from '../models';
 import { ArquivoEntity } from './../entities';
 
 @Injectable()
 export class ArquivoSerializer extends BaseSerializer<
-  ArquivoInclusaoModel,
+  ArquivoModel,
   ArquivoEntity
 > {
-  fromModel(model: ArquivoInclusaoModel): ArquivoEntity {
+  fromModel(model: ArquivoModel): ArquivoEntity {
     const entity = super.fromModel(model);
     entity.nome = model.nome;
     entity.numeroBytes = model.conteudo.size;
