+++ src/authentication/guards/authentication.guard.ts
@@ -1,6 +1,7 @@
 import {
   ExecutionContext,
   Injectable,
+  Logger,
   UnauthorizedException,
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
@@ -9,6 +10,8 @@ import { excecoes } from 'src/common/resources';
 
 @Injectable()
 export class AuthenticationGuard extends AuthGuard('jwt') {
+  private readonly logger = new Logger(AuthenticationGuard.name);
+
   constructor(private readonly reflector: Reflector) {
     super();
   }
@@ -28,7 +31,8 @@ export class AuthenticationGuard extends AuthGuard('jwt') {
 
   handleRequest(err, user) {
     if (err || !user) {
-      throw err || new UnauthorizedException(excecoes.naoAutenticado);
+      this.logger.error(err);
+      throw new UnauthorizedException(excecoes.naoAutenticado);
     }
     return user;
   }
