+++ src/common/resources/excecoes.resources.ts
@@ -1,0 +1,8 @@
+export const excecoes = {
+  tokenInvalido:
+    'JWT não possui o escopo obrigatório (`openid profile email`).',
+  naoAutenticado: 'O usuário deve estar autenticado para realizar esta ação.',
+  semAutorizacao: 'Usuário não possui autorização de acesso ao sistema.',
+  semPermissao: 'Usuário não possui autorização para a ação.',
+  erroInterno: 'Erro interno do servidor.',
+};
