+++ src/common/resources/index.ts
@@ -1,2 +1,4 @@
+export * from './excecoes.resources';
 export * from './permissoes.resources';
 export * from './strings.resources';
+export * from './validacoes.resources';
