+++ src/common/resources/validacoes.resources.ts
@@ -1,0 +1,4 @@
+export const validacoes = {
+  arquivoNaoSuportado: 'O arquivo selecionado deve estar no formato PDF.',
+  dataNaoInformada: 'A data deve ser informada.',
+};
